move official documents out to a file
